backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
greenlet==3.1.1
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
openai==1.54.4
httpx==0.27.2
nanoid==2.0.0
pdfplumber==0.11.4
python-multipart==0.0.12
pyjwt[crypto]==2.9.0
cryptography==43.0.1
icalendar==6.1.0
